backend/requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
sqlalchemy>=2.0.38
psycopg[binary]>=3.2.6
pydantic-settings>=2.8.1
python-multipart>=0.0.20
alembic>=1.14.1
pandas>=2.2.0
numpy>=1.26.0
openpyxl>=3.1.2
yfinance>=0.2.54
plotly>=5.22.0
requests>=2.32.0
PyJWT[crypto]>=2.10.1
